+++ minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lekix <lekix@student.42.fr>                +#+  +:+       +#+        */
+/*   By: kipouliq <kipouliq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 16:26:11 by kipouliq          #+#    #+#             */
-/*   Updated: 2024/05/15 14:54:37 by lekix            ###   ########.fr       */
+/*   Updated: 2024/05/16 16:36:53 by kipouliq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ typedef struct s_gbg
 typedef enum e_token_type
 {
 	CMD,
-    OPERATOR,
+	OPERATOR,
 	PIPE,
 	AND,
 	OR,
@@ -75,5 +75,13 @@ typedef enum e_gbg_lst_names
 int					gbg_coll(void *mem_addr, int which_list, int rule);
 t_env				*get_env_lst(char **envp);
 int					ft_strlen_sep(char *str, char *sep);
+t_token				*tokenize_input(char *input);
+void				print_tree(t_token **tree);
+int					ft_strlen_sep(char *str, char *sep);
+char				*find_operator(char *str);
+void				insert_operator_node(t_token **tree, t_token *node);
+void				insert_cmd_node(t_token **tree, t_token *node);
+int					ft_strcpy_sep(char *dest, char *input, char *sep);
+void				clear_tree(t_token **tree);
 
 #endif
