+++ srcs/garbage_collector.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   garbage_collector.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lekix <lekix@student.42.fr>                +#+  +:+       +#+        */
+/*   By: kipouliq <kipouliq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/09 16:04:20 by kipouliq          #+#    #+#             */
-/*   Updated: 2024/05/15 14:47:30 by lekix            ###   ########.fr       */
+/*   Updated: 2024/05/16 14:37:40 by kipouliq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	remove_mem_node(t_lst **lst, void *mem_addr)
 		return ;
 	if (ft_lstsize(*lst) == 1)
 		nullify = 1;
-	to_remove = ft_lstfind_node(lst, mem_addr);
+	to_remove = ft_lstfind_one(lst, mem_addr);
 	if (to_remove)
 	{
 		free(to_remove->content);
@@ -68,7 +68,7 @@ int	gbg_coll(void *mem_addr, int which_list, int rule)
 
 	if (rule == ADD)
 	{
-        node = ft_lstnew(mem_addr);
+		node = ft_lstnew(mem_addr);
 		if (!node)
 			return (free(mem_addr), gbg_coll(NULL, ALL, FLUSH_ALL), -1);
 		add_node_gbg(which_list, &all_gbg_lsts, node);
